requirements.txt
customtkinter==5.2.2
pillow==10.2.0
google-generativeai==0.8.3
elevenlabs==1.0.12
edge-tts==6.1.9
SpeechRecognition==3.10.1
pygame==2.5.2
pyserial==3.5
requests==2.31.0
opencv-python==4.9.0.80
python-dotenv==1.0.1
google-api-python-client==2.118.0
aiohttp==3.9.3
typing-extensions==4.9.0
matplotlib==3.8.3
pandas==2.2.0
numpy==1.26.4
qrcode==7.4.2
pyaudio==0.2.14
gTTS==2.5.1
